backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
drf-spectacular==0.28.0
django-cors-headers==4.9.0
python-dotenv==1.1.1
django-filter==25.1
djangorestframework-simplejwt==5.5.1
